Use context and lazy components

+++ frontend/src/components/ACLPage/List/List.tsx
@@ -1,5 +1,5 @@
-import React, { useState } from 'react';
-import { ColumnDef } from '@tanstack/react-table';
+import React from 'react';
+import { ColumnDef, Row } from '@tanstack/react-table';
 import PageHeading from 'components/common/PageHeading/PageHeading';
 import Table from 'components/common/NewTable';
 import { useConfirm } from 'lib/hooks/useConfirm';
@@ -16,33 +16,35 @@ import { Button } from 'components/common/Button/Button';
 import ACLForm from 'components/ACLPage/Form/Form';
 
 import * as S from './List.styled';
-import ActionsCell from './ActionsCell';
+import DeleteIcon from 'components/common/Icons/DeleteIcon';
+import { useTheme } from 'styled-components';
+import ACLFormContext from '../Form/AclFormContext';
 
 const ACList: React.FC = () => {
   const { clusterName } = useAppParams<{ clusterName: ClusterName }>();
   const { data: aclList } = useAcls(clusterName);
   const { deleteResource } = useDeleteAcl(clusterName);
   const modal = useConfirm(true);
+  const theme = useTheme();
   const {
     value: isFormOpen,
     setFalse: closeForm,
     setTrue: openFrom,
   } = useBoolean();
+  const [rowId, setRowId] = React.useState('');
 
-  const handleDeleteClick = (acl: KafkaAcl) => {
-    modal('Are you sure want to delete this ACL record?', () =>
-      deleteResource(acl)
-    );
+  const handleDeleteClick = (acl: KafkaAcl | null) => {
+    if (acl) {
+      modal('Are you sure want to delete this ACL record?', () =>
+        deleteResource(acl)
+      );
+    }
   };
-  const [currentAcl, setAcl] = useState<KafkaAcl | null>(null);
 
-  const handleEditClick = (acl: KafkaAcl) => {
-    setAcl(acl);
-    openFrom();
-  };
-  const handleCreateClick = () => {
-    setAcl(null);
-    openFrom();
+  const handleRowHover = (value: Row<KafkaAcl>) => {
+    if (value) {
+      setRowId(value.id);
+    }
   };
 
   const columns = React.useMemo<ColumnDef<KafkaAcl>[]>(
@@ -123,16 +125,20 @@ const ACList: React.FC = () => {
         size: 111,
       },
       {
-        id: 'actions',
+        id: 'delete',
         // eslint-disable-next-line react/no-unstable-nested-components
         cell: ({ row }) => {
           return (
-            <ActionsCell
-              onDelete={() => handleDeleteClick(row.original)}
-              onEdit={() => handleEditClick(row.original)}
-            />
+            <S.DeleteCell onClick={() => handleDeleteClick(row.original)}>
+              <DeleteIcon
+                fill={
+                  rowId === row.id ? theme.acl.table.deleteIcon : 'transparent'
+                }
+              />
+            </S.DeleteCell>
           );
         },
+        size: 76,
       },
     ],
     []
@@ -141,7 +147,7 @@ const ACList: React.FC = () => {
   return (
     <S.Container>
       <PageHeading text="Access Control List">
-        <Button buttonType="primary" buttonSize="M" onClick={handleCreateClick}>
+        <Button buttonType="primary" buttonSize="M" onClick={openFrom}>
           + Create ACL
         </Button>
       </PageHeading>
@@ -149,8 +155,12 @@ const ACList: React.FC = () => {
         columns={columns}
         data={aclList ?? []}
         emptyMessage="No ACL items found"
+        onRowHover={handleRowHover}
+        onMouseLeave={() => setRowId('')}
       />
-      <ACLForm open={isFormOpen} onClose={closeForm} acl={currentAcl} />
+      <ACLFormContext.Provider value={{ onClose: closeForm }}>
+        {isFormOpen && <ACLForm isOpen={isFormOpen} />}
+      </ACLFormContext.Provider>
     </S.Container>
   );
 };
